IMTEPD-210: Added getNumDocs() to the ImhotepSession api

+++ imhotep-server/src/main/java/com/indeed/imhotep/local/IndexReWriter.java
@@ -176,7 +176,7 @@ public class IndexReWriter {
 
         for (int i = 0; i < this.sessions.size(); i++) {
             int offset = this.sessionDocIdOffsets[i];
-            int nDocs = this.sessions.get(i).getNumDocs();
+            int nDocs = (int)this.sessions.get(i).getNumDocs();
             if (!memory.claimMemory(nDocs * 4L))
                 throw new ImhotepOutOfMemoryException();
             int[] mapping = new int[nDocs];
